+++ CHANGELOG.md
@@ -7,6 +7,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Post-install script support.
+- Tooth path validation.
+- Flexible tooth.json parsing.
+
+## [0.3.4] - 2023-01-25
+
+### Changed
+
+- Bumped github.com/fatih/color from 1.14.0 to 1.14.1.
+
+### Fixed
+
+- Misleading error hints.
+- Failing to fetch tooth with major version v0 or v1.
+- Failing to match dependencies.
+- Failing to fetch tooth when uppercase letters exist in tooth path.
+
+## [0.3.3] - 2023-01-24
+
+### Fixed
+
+- Default to earliest version when no version is specified in tooth.json.
+- Panic when tooth.json is invalid.
+
+## [0.3.2] - 2023-01-23
+
+### Added
+
+- "Add to PATH" option in setup utility.
+- Mac OS, Linux and OpenBSD support.
+- Arm64 support.
+
+## [0.3.1] - 2023-01-21
+
+### Added
+
+- Setup utility to install Lip.
+
 ## [0.3.0] - 2023-01-20
 
 ### Added
@@ -43,7 +83,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Basic functions: cache, install, list, show, tooth init, and uninstall.
 
-[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/LiteLDev/Lip/compare/v0.3.4...HEAD
+[0.3.4]: https://github.com/LiteLDev/Lip/releases/tag/v0.3.3...v0.3.4
+[0.3.3]: https://github.com/LiteLDev/Lip/releases/tag/v0.3.2...v0.3.3
+[0.3.2]: https://github.com/LiteLDev/Lip/releases/tag/v0.3.1...v0.3.2
+[0.3.1]: https://github.com/LiteLDev/Lip/releases/tag/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/LiteLDev/Lip/releases/tag/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/LiteLDev/Lip/releases/tag/v0.1.0...v0.2.0